move scripts to bin/ directory
